# output/src/ApiApp/Controllers/OrdersController.cs
// Generated by booster-apidev | 2026-04-06 19:37:30

using ApiApp.Data;
using ApiApp.DTOs.OrderDetails;
using ApiApp.DTOs.Orders;
using ApiApp.Models;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;

namespace ApiApp.Controllers;

[ApiController]
[Route("api/orders")]
public class OrdersController : ControllerBase
{
    private readonly AppDbContext _db;

    public OrdersController(AppDbContext db)
    {
        _db = db;
    }

    // GET /api/orders
    [HttpGet]
    public async Task<ActionResult<DTOs.PagedResult<OrderListItemDto>>> GetOrders(
        [FromQuery] int? customerId,
        [FromQuery] int? employeeId,
        [FromQuery] string? status,
        [FromQuery] string? sortBy,
        [FromQuery] int? top,
        [FromQuery] int page = 1,
        [FromQuery] int pageSize = 20)
    {
        var query = _db.Orders
            .Include(o => o.Customer)
            .Include(o => o.Employee)
            .Include(o => o.OrderDetails)
            .AsQueryable();

        if (customerId.HasValue) query = query.Where(o => o.CustomerId == customerId);
        if (employeeId.HasValue) query = query.Where(o => o.EmployeeId == employeeId);

        if (!string.IsNullOrWhiteSpace(status) && Enum.TryParse<OrderStatus>(status, true, out var statusEnum))
        {
            query = query.Where(o => o.Status == statusEnum);
        }

        query = sortBy?.ToLower() == "modifiedon"
            ? query.OrderByDescending(o => o.ModifiedOn)
            : query.OrderByDescending(o => o.OrderDate);

        if (top.HasValue && top.Value > 0)
        {
            var topItems = await query.Take(top.Value).ToListAsync();
            return Ok(new DTOs.PagedResult<OrderListItemDto>
            {
                Items = topItems.Select(MapToListItem),
                TotalCount = topItems.Count,
                Page = 1,
                PageSize = top.Value
            });
        }

        pageSize = Math.Min(pageSize, 100);
        page = Math.Max(page, 1);
        var totalCount = await query.CountAsync();
        var items = await query.Skip((page - 1) * pageSize).Take(pageSize).ToListAsync();

        return Ok(new DTOs.PagedResult<OrderListItemDto>
        {
            Items = items.Select(MapToListItem),
            TotalCount = totalCount,
            Page = page,
            PageSize = pageSize
        });
    }

    // GET /api/orders/{id}
    [HttpGet("{id:int}")]
    public async Task<ActionResult<OrderDetailDto>> GetOrder(int id)
    {
        var order = await _db.Orders
            .Include(o => o.Customer)
            .Include(o => o.Employee)
            .Include(o => o.OrderDetails)
                .ThenInclude(od => od.Product)
            .FirstOrDefaultAsync(o => o.Id == id);

        if (order == null) return NotFound(new { error = $"Order {id} not found." });
        return Ok(MapToDetail(order));
    }

    // POST /api/orders
    [HttpPost]
    public async Task<ActionResult<OrderDetailDto>> CreateOrder([FromBody] OrderCreateDto dto)
    {
        OrderStatus status = OrderStatus.New;
        if (!string.IsNullOrWhiteSpace(dto.Status) && !Enum.TryParse(dto.Status, true, out status))
        {
            return BadRequest(new { errors = new { Status = new[] { $"Invalid status value: {dto.Status}" } } });
        }

        var order = new Order
        {
            CustomerId = dto.CustomerId,
            EmployeeId = dto.EmployeeId,
            OrderDate = dto.OrderDate,
            ShippedDate = dto.ShippedDate,
            PaidDate = dto.PaidDate,
            Status = status,
            Notes = dto.Notes
        };

        _db.Orders.Add(order);
        await _db.SaveChangesAsync();

        // Reload with navigations
        await _db.Entry(order).Reference(o => o.Customer).LoadAsync();
        await _db.Entry(order).Reference(o => o.Employee).LoadAsync();
        await _db.Entry(order).Collection(o => o.OrderDetails).LoadAsync();

        return CreatedAtAction(nameof(GetOrder), new { id = order.Id }, MapToDetail(order));
    }

    // PUT /api/orders/{id}
    [HttpPut("{id:int}")]
    public async Task<ActionResult<OrderDetailDto>> UpdateOrder(int id, [FromBody] OrderCreateDto dto)
    {
        var order = await _db.Orders
            .Include(o => o.Customer)
            .Include(o => o.Employee)
            .Include(o => o.OrderDetails)
                .ThenInclude(od => od.Product)
            .FirstOrDefaultAsync(o => o.Id == id);

        if (order == null) return NotFound(new { error = $"Order {id} not found." });

        OrderStatus status = OrderStatus.New;
        if (!string.IsNullOrWhiteSpace(dto.Status) && !Enum.TryParse(dto.Status, true, out status))
        {
            return BadRequest(new { errors = new { Status = new[] { $"Invalid status value: {dto.Status}" } } });
        }

        order.CustomerId = dto.CustomerId;
        order.EmployeeId = dto.EmployeeId;
        order.OrderDate = dto.OrderDate;
        order.ShippedDate = dto.ShippedDate;
        order.PaidDate = dto.PaidDate;
        order.Status = status;
        order.Notes = dto.Notes;

        await _db.SaveChangesAsync();
        return Ok(MapToDetail(order));
    }

    // PATCH /api/orders/{id}/status
    [HttpPatch("{id:int}/status")]
    public async Task<ActionResult<OrderStatusPatchResponseDto>> PatchOrderStatus(int id, [FromBody] OrderStatusPatchDto dto)
    {
        if (!Enum.TryParse<OrderStatus>(dto.Status, true, out var newStatus))
        {
            return BadRequest(new { errors = new { Status = new[] { $"Invalid status value: {dto.Status}" } } });
        }

        var order = await _db.Orders.FindAsync(id);
        if (order == null) return NotFound(new { error = $"Order {id} not found." });

        order.Status = newStatus;
        await _db.SaveChangesAsync();

        return Ok(new OrderStatusPatchResponseDto
        {
            Id = order.Id,
            Status = order.Status.ToString(),
            StatusDisplay = order.Status.ToString(),
            ModifiedOn = order.ModifiedOn
        });
    }

    // DELETE /api/orders/{id}
    [HttpDelete("{id:int}")]
    public async Task<IActionResult> DeleteOrder(int id)
    {
        var order = await _db.Orders.FindAsync(id);
        if (order == null) return NotFound(new { error = $"Order {id} not found." });

        _db.Orders.Remove(order); // cascade deletes OrderDetails via EF config
        await _db.SaveChangesAsync();
        return NoContent();
    }

    private static OrderListItemDto MapToListItem(Order o) => new()
    {
        Id = o.Id,
        CustomerId = o.CustomerId,
        CustomerName = o.Customer?.CustomerName,
        EmployeeId = o.EmployeeId,
        EmployeeFullName = o.Employee != null
            ? $"{o.Employee.FirstName} {o.Employee.LastName}".Trim()
            : null,
        OrderDate = o.OrderDate,
        ShippedDate = o.ShippedDate,
        PaidDate = o.PaidDate,
        Status = o.Status.ToString(),
        StatusDisplay = o.Status.ToString(),
        OrderTotal = o.OrderTotal,
        ModifiedOn = o.ModifiedOn
    };

    internal static OrderDetailDto MapToDetail(Order o) => new()
    {
        Id = o.Id,
        CustomerId = o.CustomerId,
        CustomerName = o.Customer?.CustomerName,
        EmployeeId = o.EmployeeId,
        EmployeeFullName = o.Employee != null
            ? $"{o.Employee.FirstName} {o.Employee.LastName}".Trim()
            : null,
        OrderDate = o.OrderDate,
        ShippedDate = o.ShippedDate,
        PaidDate = o.PaidDate,
        Status = o.Status.ToString(),
        StatusDisplay = o.Status.ToString(),
        Notes = o.Notes,
        OrderTotal = o.OrderTotal,
        OrderDetails = o.OrderDetails.Select(od => new OrderDetailItemDto
        {
            Id = od.Id,
            OrderId = od.OrderId,
            ProductId = od.ProductId,
            ProductCode = od.Product?.ProductCode,
            ProductName = od.Product?.ProductName,
            Quantity = od.Quantity,
            UnitPrice = od.UnitPrice,
            ExtendedPrice = od.ExtendedPrice
        }).ToList(),
        AddedBy = o.AddedBy,
        AddedOn = o.AddedOn,
        ModifiedBy = o.ModifiedBy,
        ModifiedOn = o.ModifiedOn
    };
}